account/requirements.txt
Flask
setuptools
paramiko
fabric3